+++ scripts/macros/classFeatures/artificer/infusions/enhancedWeapon.js
@@ -31,6 +31,8 @@ async function attack({speaker, actor, token, character, item, args, scope, work
     attackBonus = level;
     properties.mgc = true;
     workflow.item = workflow.item.clone({'system.damage.parts': parts, 'system.properties': properties, 'system.attackBonus': attackBonus}, {'keepId': true});
+    workflow.item.prepareData();
+    workflow.item.prepareFinalAttributes();
     queue.remove(workflow.item.uuid);
 }
 export let enhancedWeapon = {