+++ scripts/macros/homebrew/bg3/concussiveSmash.js
@@ -4,7 +4,11 @@ import {constants} from '../../../constants.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
     let queueSetup = await queue.setup(workflow.item.uuid, 'concussiveSmash', 50);
     if (!queueSetup) return;
-    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) {
+        workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+        workflow.item.prepareData();
+        workflow.item.prepareFinalAttributes();
+    }
     queue.remove(workflow.item.uuid);
 }
 async function save({speaker, actor, token, character, item, args, scope, workflow}) {
@@ -72,7 +76,7 @@ async function attack({speaker, actor, token, character, item, args, scope, work
     if (!feature.system.uses.value) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'concussiveSmash', 250);
     if (!queueSetup) return;
-    let selection = await chris.dialog(feature.name, [['Yes', true], ['No', false]], 'Use Concussive Smash?');
+    let selection = await chris.dialog(feature.name, constants.yesNo, 'Use Concussive Smash?');
     if (!selection) {
         queue.remove(workflow.item.uuid);
         return;