+++ scripts/macros/homebrew/bg3/flourish.js
@@ -1,5 +1,6 @@
 import {chris} from '../../../helperFunctions.js';
-export async function flourish(effect) {
+import {queue} from '../../../utility/queue.js';
+async function effect(effect) {
     if (effect.flags['chris-premades']?.feature?.flourish) {
         await chris.removeEffect(effect);
         return;
@@ -8,4 +9,18 @@ export async function flourish(effect) {
         'flags.chris-premades.feature.flourish': true
     };
     await chris.updateEffect(effect, updates);
+}
+async function item({speaker, actor, token, character, item, args, scope, workflow}) {
+    let queueSetup = await queue.setup(workflow.item.uuid, 'flourish', 50);
+    if (!queueSetup) return;
+    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) {
+        workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+        workflow.item.prepareData();
+        workflow.item.prepareFinalAttributes();
+    }
+    queue.remove(workflow.item.uuid);
+}
+export let flourish = {
+    'effect': effect,
+    'item': item
 }