+++ scripts/macros/homebrew/bg3/lacerate.js
@@ -4,7 +4,11 @@ import {constants} from '../../../constants.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
     let queueSetup = await queue.setup(workflow.item.uuid, 'lacerate', 50);
     if (!queueSetup) return;
-    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) {
+        workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+        workflow.item.prepareData();
+        workflow.item.prepareFinalAttributes();
+    }
     queue.remove(workflow.item.uuid);
     if (workflow.targets.size != 1) return;
     let race = chris.raceOrType(workflow.targets.first().actor);
@@ -73,7 +77,7 @@ async function attack({speaker, actor, token, character, item, args, scope, work
     if (!feature.system.uses.value) return;
     let queueSetup = await queue.setup(workflow.item.uuid, 'lacerate', 250);
     if (!queueSetup) return;
-    let selection = await chris.dialog(feature.name, [['Yes', true], ['No', false]], 'Use Lacerate?');
+    let selection = await chris.dialog(feature.name, constants.yesNo, 'Use Lacerate?');
     if (!selection) {
         queue.remove(workflow.item.uuid);
         return;