+++ scripts/macros/homebrew/bg3/rushAttack.js
@@ -3,7 +3,11 @@ import {queue} from '../../../utility/queue.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
     let queueSetup = await queue.setup(workflow.item.uuid, 'rushAttack', 50);
     if (!queueSetup) return;
-    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+    if (workflow.actor.system.abilities.dex.save > workflow.actor.system.abilities.str.save) {
+        workflow.item = workflow.item.clone({'system.save.scaling': 'dex'}, {'keepId': true});
+        workflow.item.prepareData();
+        workflow.item.prepareFinalAttributes();
+    }
     queue.remove(workflow.item.uuid);
 }
 async function turn(effect) {